added fragment for historical tests
