+++ app/src/main/java/org/inspire/breath/fragments/home/History.java
@@ -1,4 +1,25 @@
 package org.inspire.breath.fragments.home;
 
-public class History {
+import android.os.Bundle;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import org.inspire.breath.R;
+import org.inspire.breath.utils.FragmentedFragment;
+
+public class History extends FragmentedFragment {
+
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_history, container, false);
+    }
+
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+    }
 }
